+++ src/IntelliTect.Coalesce.CodeGeneration.Vue/Generators/Scripts/TsViewModels.cs
@@ -19,7 +19,7 @@ public class TsViewModels : StringBuilderFileGenerator<ReflectionRepository>
         b.Line("import * as $metadata from './metadata.g'");
         b.Line("import * as $models from './models.g'");
         b.Line("import * as $apiClients from './api-clients.g'");
-        b.Line("import { ViewModel, ListViewModel, ViewModelCollection, ServiceViewModel, type DeepPartial, defineProps, createAbstractProxyViewModelType } from 'coalesce-vue/lib/viewmodel'");
+        b.Line("import { ViewModel, ListViewModel, ViewModelCollection, ServiceViewModel, type DeepPartial, defineProps, createAbstractLoader } from 'coalesce-vue/lib/viewmodel'");
         b.Line();
 
         foreach (var model in Model.CrudApiBackedClasses.OrderBy(e => e.ClientTypeName))
@@ -80,7 +80,14 @@ public class TsViewModels : StringBuilderFileGenerator<ReflectionRepository>
         if (model.Type.IsAbstract)
         {
             b.Line($"export type {viewModelName} = {string.Join(" | ", model.ClientDerivedTypes.Select(t => new VueType(t.Type).TsType(viewModel: true)))}");
-            b.Line($"export const {viewModelName} = createAbstractProxyViewModelType<{modelName}, {viewModelName}>({metadataName}, $apiClients.{name}ApiClient)");
+            if (model.ClientDataSources(Model).Any())
+            {
+                b.Line($"export const {viewModelName} = createAbstractLoader<{viewModelName}>($apiClients.{name}ApiClient, {{ DataSources: {modelName}.DataSources }})");
+            }
+            else
+            {
+                b.Line($"export const {viewModelName} = createAbstractLoader<{viewModelName}>($apiClients.{name}ApiClient)");
+            }
             b.Line();
             return;
         }