+++ src/coalesce-vue/test/api-client.spec.ts
@@ -684,7 +684,7 @@ describe("$makeCaller", () => {
     );
 
     const arg = 42;
-    const result = await caller(arg);
+    const result: number = await caller(arg);
     expect(endpointMock.mock.calls[0][0]).toBe(arg);
     expect(caller.result).toBe(arg);
     expect(result).toBe(arg);
@@ -727,15 +727,16 @@ describe("$makeCaller", () => {
     const arg = 42;
     const result = await caller(arg);
 
-    // The typings are actually wrong at the moment - `undefined` is not one of the types of `result`, but it should be.
+    // `result` can be undefined since `undefined` is part of the return signature of the invoker func.
+    const _resultAllowsNumber: typeof result = 24;
+    const _resultAllowsUndefined: typeof result = undefined;
+
     expect(result).toBeUndefined();
 
     expect(endpointMock.mock.calls.length).toBe(0);
     expect(caller.result).toBeNull();
 
     // Typescript typing tests - all of these are valid types of `result`.
-    // Note that Typescript intellisense in VS code seems to be really messed up
-    // right now and shows that `result` is only `string`.:
     caller.result = null;
     caller.result = 13;
 
@@ -759,15 +760,16 @@ describe("$makeCaller", () => {
     const arg = 42;
     const result = await caller(arg);
 
-    // The typings are actually wrong at the moment - `undefined` is not one of the types of `result`, but it should be.
+    // `result` can be undefined since `undefined` is part of the return signature of the invoker func.
+    const _resultAllowsNumber: typeof result = 24;
+    const _resultAllowsUndefined: typeof result = undefined;
+
     expect(result).toBeUndefined();
 
     expect(endpointMock.mock.calls.length).toBe(0);
     expect(caller.result).toBeNull();
 
     // Typescript typing tests - all of these are valid types of `result`.
-    // Note that Typescript intellisense in VS code seems to be really messed up
-    // right now and shows that `result` is only `string`.:
     caller.result = null;
     caller.result = 13;
 
@@ -1437,15 +1439,16 @@ describe("$makeCaller with args object", () => {
     const arg = 42;
     const result = await caller.invokeWithArgs({ num: arg });
 
-    // The typings are actually wrong at the moment - `undefined` is not one of the types of `result`, but it should be.
+    // `result` can be undefined since `undefined` is part of the return signature of the invoker func.
+    const _resultAllowsNumber: typeof result = 24;
+    const _resultAllowsUndefined: typeof result = undefined;
+
     expect(result).toBeUndefined();
 
     expect(endpointMock.mock.calls.length).toBe(0);
     expect(caller.result).toBeNull();
 
     // Typescript typing tests - all of these are valid types of `result`.
-    // Note that Typescript intellisense in VS code seems to be really messed up
-    // right now and shows that `result` is only `string`.:
     caller.result = null;
     caller.result = 13;
 
@@ -1500,6 +1503,15 @@ describe("$makeCaller with args object", () => {
     const _pageSize: number | null = caller.pageSize;
     const _totalCount: number | null = caller.totalCount;
 
+    () => {
+      //@ts-expect-error list should not allow `any[]`
+      caller.rawResponse.data.list = [Symbol()];
+      //@ts-expect-error list should not allow any
+      caller.rawResponse.data.list = Symbol();
+      // Should allow numbers
+      caller.rawResponse.data.list = [42, 67];
+    };
+
     const promiseResult: number[] = await caller(42);
     expect(promiseResult).toStrictEqual([42]);
     expect(caller.result).toStrictEqual([42]);
@@ -1516,11 +1528,28 @@ describe("$makeCaller with args object", () => {
   });
 
   describe.each(["item", "list"] as const)("for %s transport", (type) => {
-    const makeCaller = (
-      endpointMock: ReturnType<
-        typeof makeEndpointMock<number | null | undefined>
-      >,
-    ) =>
+    const makeTransportMock = () =>
+      vitest.fn((arg?: number | null | undefined) =>
+        Promise.resolve({
+          data:
+            type === "list"
+              ? {
+                  wasSuccessful: true,
+                  list: [arg],
+                  page: 1,
+                  pageCount: 1,
+                  pageSize: 10,
+                  totalCount: 1,
+                }
+              : { wasSuccessful: true, object: arg },
+          status: 200,
+          statusText: "OK",
+          headers: {},
+          config: {} as any,
+        }),
+      );
+
+    const makeCaller = (endpointMock: ReturnType<typeof makeTransportMock>) =>
       new PersonApiClient().$makeCaller(
         type,
         (c, num: number) => endpointMock(num),
@@ -1529,7 +1558,7 @@ describe("$makeCaller with args object", () => {
       );
 
     test("uses own args if args not specified", () => {
-      const endpointMock = makeEndpointMock();
+      const endpointMock = makeTransportMock();
       const caller = makeCaller(endpointMock);
 
       caller.args.num = 42;
@@ -1538,7 +1567,7 @@ describe("$makeCaller with args object", () => {
     });
 
     test("own args are reactive", async () => {
-      const endpointMock = makeEndpointMock();
+      const endpointMock = makeTransportMock();
       const caller = makeCaller(endpointMock);
 
       const vue = mountData({ caller });
@@ -1558,7 +1587,7 @@ describe("$makeCaller with args object", () => {
     });
 
     test("uses custom args if specified", () => {
-      const endpointMock = makeEndpointMock();
+      const endpointMock = makeTransportMock();
       const caller = makeCaller(endpointMock);
 
       caller.args.num = 42;
@@ -1569,7 +1598,7 @@ describe("$makeCaller with args object", () => {
     });
 
     test("sets state properties appropriately", async () => {
-      const endpointMock = makeEndpointMock();
+      const endpointMock = makeTransportMock();
       const caller = makeCaller(endpointMock);
 
       caller.args.num = 42;
@@ -1581,7 +1610,7 @@ describe("$makeCaller with args object", () => {
     });
 
     test("debounce ignores redundant requests when resolving", async () => {
-      const endpointMock = makeEndpointMock();
+      const endpointMock = makeTransportMock();
       const caller = new PersonApiClient()
         .$makeCaller(
           type,
@@ -1613,7 +1642,7 @@ describe("$makeCaller with args object", () => {
       expect(endpointMock.mock.calls[1][0]).toBe(3);
 
       await expect(calls[0]).resolves.toBeTruthy();
-      await expect(calls[1]).resolves.toBeFalsy();
+      await expect(calls[1]).resolves.toBeUndefined();
       await expect(calls[2]).resolves.toBeTruthy();
     });
   });